Attempt 2 - spec names and set_position
This attempt rectifies two of the things I liked least about my first
attempt - the spec descriptions and the set_position_of method.

The spec descriptions are a lot better, and set_position_of has been
made private. There should be no way to set the position of a player in
the normal course of the game outside the roll() method - it shouldn't
be part of the API. Instead, the positions of players are passed in
through the constructor. To allow this to be set in the Givens of the
scenario, I had to create a BoardBuilder which would take options and
would construct a board when it was first needed in the scenarios.